exclude currently active user from user removal
When installing a new-installed server with basic installation, the primary user (Ubuntu: `ubuntu`, Raspberry Pi: `pi`, ...) may not be deleted ... so for the playbook run not failing, the `ansible_ssh_user` is also excluded from removal.

For final installation, the playbook `playbooks/utils/rollout_users` has to be run with real final connection user!